requirements.txt
asgiref==3.3.1
CacheControl==0.12.6
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cloudinary==1.24.0
cryptography==3.3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-braces==1.14.0
django-filter==2.4.0
django-firebase-orm==0.6.1
django-friendship==1.9.1
django-heroku==0.3.1
django-location-field==2.1.0
django-model-utils==4.1.1
django-phonenumber-field==5.0.0
django-star-ratings==0.9.1
firebase-admin==2.13.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
msgpack==1.0.2
oauthlib==3.1.0
phonenumbers==8.12.15
proto-plus==1.13.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
